rename input_formatter module
