+++ benchmark/dbally_benchmark/iql_benchmark.py
@@ -23,7 +23,7 @@ from dbally.audit.event_tracker import EventTracker
 from dbally.iql_generator.iql_generator import IQLGenerator
 from dbally.iql_generator.iql_prompt_template import UnsupportedQueryError, default_iql_template
 from dbally.llms.litellm import LiteLLM
-from dbally.prompts.input_format import DefaultInputFormatter
+from dbally.prompts.formatters import DefaultInputFormatter
 from dbally.views.structured import BaseStructuredView
