+++ tests/unit/test_iql_format.py
@@ -4,7 +4,7 @@ import pytest
 
 from dbally.iql_generator.iql_prompt_template import default_iql_template
 from dbally.prompts.few_shot import FewShotExample
-from dbally.prompts.input_format import DefaultFewShotInputFormatter, DefaultInputFormatter
+from dbally.prompts.formatters import DefaultFewShotInputFormatter, DefaultInputFormatter
 
 
 async def test_iql_input_format_default() -> None:
